fix status bar iOS 6

+++ SmartGuide/SmartGuide/GUI/ViewControllers/SearchShop/ShopUser/ShopKM/KM2/SUKM2Cell.m
@@ -67,7 +67,7 @@
 -(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
     KM2Voucher *voucher=_vouchers[indexPath.row];
-    return voucher.voucherHeight;
+    return voucher.voucherHeight.floatValue;
 }
 
 -(UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
